fixing bug: can't change a task of status once was created
