+++ src/app/components/task/list/list-task.component.html
@@ -1,27 +1,27 @@
 <!-- Crear componente columnas status -->
 <div class="flex min-h-screen">
 
-  <div (dragover)="allowDrop($event)" (drop)="task.dropPending()"
+  <div (dragover)="allowDrop($event)" (drop)="dropPending()"
     class="drag-column drop-column shadow-1 bg-cyan-50 col-4">
 
     <h3 class="white-space-normal">Pending <i class="pi pi-clock"></i></h3>
-    <app-task-item #task (loadAll)="loadAllTasks()" [tasks]="pendingTasks"></app-task-item>
+    <app-task-item (draggedStart)="dragStart($event)" (loadAll)="loadAllTasks()" [tasks]="pendingTasks"></app-task-item>
 
   </div>
 
-  <div (dragover)="allowDrop($event)" (drop)="task.dropDoing()"
+  <div (dragover)="allowDrop($event)" (drop)="dropDoing()"
     class="drag-column drop-column shadow-1 card surface-100 col-4">
 
     <h3 class="white-space-normal">Doing <i class="pi pi-chart-line"></i></h3>
-    <app-task-item (loadAll)="loadAllTasks()" [tasks]="doingTasks"></app-task-item>
+    <app-task-item (draggedStart)="dragStart($event)" (loadAll)="loadAllTasks()" [tasks]="doingTasks"></app-task-item>
 
   </div>
 
-  <div (dragover)="allowDrop($event)" (drop)="task.dropDone()"
+  <div (dragover)="allowDrop($event)" (drop)="dropDone()"
     class="drag-column drop-column shadow-1 card bg-teal-100 col-4">
 
     <h3 class="white-space-normal">Done <i class="pi pi-check-square"></i></h3>
-    <app-task-item (loadAll)="loadAllTasks()"  [tasks]="doneTasks"></app-task-item>
+    <app-task-item (draggedStart)="dragStart($event)" (loadAll)="loadAllTasks()"  [tasks]="doneTasks"></app-task-item>
 
   </div>
