+++ src/app/components/task/list/list-task.component.ts
@@ -1,6 +1,7 @@
 import {  Component, OnInit } from '@angular/core';
 import { TaskService } from 'src/app/core/services/task.service';
 import { Task } from 'src/app/core/models/task.model';
+import { TaskStatus } from 'src/app/core/enums/task-status.enum';
 
 
 @Component({
@@ -12,6 +13,7 @@ export class ListTaskComponent implements OnInit {
   pendingTasks!: Task[];
   doingTasks!: Task[];
   doneTasks!: Task[];
+  draggedTask: Task = null;
 
   constructor(private taskService: TaskService) {}
 
@@ -28,5 +30,34 @@ export class ListTaskComponent implements OnInit {
   allowDrop($event: DragEvent): void{
    $event.preventDefault();
  }
+
+ public dropPending() {
+
+  this.draggedTask.taskStatus = TaskStatus.Pending;
+  this.taskService.saveTask(this.draggedTask);
+  this.loadAllTasks();
+}
+
+public dropDoing(){
+
+  this.draggedTask.taskStatus = TaskStatus.Doing;
+  this.taskService.saveTask(this.draggedTask);
+  this.loadAllTasks();
+}
+
+public dropDone(){
+
+  this.draggedTask.taskStatus = TaskStatus.Done;
+  this.taskService.saveTask(this.draggedTask);
+  this.loadAllTasks();
+}
+
+dragStart(task: Task) {
+  this.draggedTask = task;
+}
+
+dragEnd() {
+  this.draggedTask = null;
+}
 
 }
