+++ src/app/components/task/task-item/task-item.component.ts
@@ -11,41 +11,17 @@ import { TaskService } from 'src/app/core/services/task.service';
 export class TaskItemComponent implements OnInit {
 
   @Input() tasks: Task[];
-  @Output() loadAll = new EventEmitter();
+  @Output() draggedStart = new EventEmitter();
   draggedTask: Task = null;
 
-  constructor(private taskService: TaskService) { }
+  constructor() { }
 
   ngOnInit(): void {
   }
 
-  dragStart(task: Task) {
-    this.draggedTask = task;
+  public startDragging($event: Task){
+    this.draggedStart.emit($event);
   }
 
-  dragEnd() {
-    this.draggedTask = null;
-  }
-
-  public dropPending() {
-
-    this.draggedTask.taskStatus = TaskStatus.Pending;
-    this.taskService.saveTask(this.draggedTask);
-    this.loadAll.emit();
-  }
-
-  public dropDoing(){
-
-    this.draggedTask.taskStatus = TaskStatus.Doing;
-    this.taskService.saveTask(this.draggedTask);
-    this.loadAll.emit();
-  }
-
-  public dropDone(){
-
-    this.draggedTask.taskStatus = TaskStatus.Done;
-    this.taskService.saveTask(this.draggedTask);
-    this.loadAll.emit();
-  }
 
 }
